Adding Record.get_stream()
Record.get_stream() will return a file-like object with support for the
read() method.  Calls to read() on this object will be passed to the
underlying source layer; this can be used to process a large file while
it is still being read from a slow disk or the network.